refactor: move ssh importer errors to sub package.
So that external packages can grok the errors being produced by the ssh
importer without having to take on the package dependencies. This was
introduced for the keymanager service.
