+++ internal/ssh/importer/errors/errors.go
@@ -1,7 +1,7 @@
 // Copyright 2024 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package importer
+package errors
 
 import (
 	"github.com/juju/errors"
