+++ internal/ssh/importer/launchpad.go
@@ -11,6 +11,8 @@ import (
 	"net/url"
 
 	"github.com/juju/errors"
+
+	importererrors "github.com/juju/juju/internal/ssh/importer/errors"
 )
 
 // LaunchpadResolver is an implementation of [Resolver] for retrieving the
@@ -33,8 +35,8 @@ const (
 // Launchpad subject in this case a user and returning all of the public ssh
 // keys the user has for their profile.
 // The following errors can be expected:
-// - [SubjectNotFound] when the subject being asked for does not exist in
-// the resolvers domain.
+// - [importererrors.SubjectNotFound] when the subject being asked for does not
+// exist in the resolvers domain.
 func (l *LaunchpadResolver) PublicKeysForSubject(
 	ctx context.Context,
 	subject string,
@@ -66,7 +68,7 @@ func (l *LaunchpadResolver) PublicKeysForSubject(
 	if res.StatusCode == http.StatusNotFound {
 		return nil, fmt.Errorf(
 			"cannot find launchpad user %q%w",
-			subject, errors.Hide(SubjectNotFound),
+			subject, errors.Hide(importererrors.SubjectNotFound),
 		)
 	} else if res.StatusCode != http.StatusOK {
 		return nil, fmt.Errorf(
